* fix tests after merge.
