+++ laws/src/main/scala/cats/laws/discipline/ArbitraryK.scala
@@ -3,6 +3,7 @@ package cats.laws.discipline
 import cats.data.{Cokleisli, Kleisli, Or, Const}
 import org.scalacheck.Arbitrary
 import cats.laws.discipline.arbitrary._
+import scala.concurrent.{ExecutionContext, Future}
 
 trait ArbitraryK[F[_]] {
   def synthesize[A: Arbitrary]: Arbitrary[F[A]]
@@ -41,4 +42,10 @@ object ArbitraryK {
 
   implicit def cokleisliA[F[_], A]: ArbitraryK[Cokleisli[F, A, ?]] =
     new ArbitraryK[Cokleisli[F, A, ?]]{ def synthesize[B: Arbitrary]: Arbitrary[Cokleisli[F, A, B]] = implicitly }
+
+  implicit def futureArbitraryK(implicit ec: ExecutionContext): ArbitraryK[Future] =
+    new ArbitraryK[Future] {
+      def synthesize[A](implicit A: Arbitrary[A]): Arbitrary[Future[A]] =
+        Arbitrary(A.arbitrary.map(Future.successful))
+    }
 }
